requirements.txt
allure-pytest
allure-python-commons
pytest
pytest-xdist
selenium